+++ src/ext-test/java/org/opentripplanner/ext/vectortiles/layers/LayerFiltersTest.java
@@ -6,7 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.time.LocalDate;
 import java.util.List;
 import org.junit.jupiter.api.Test;
-import org.opentripplanner.apis.gtfs.PatternTestModel;
+import org.opentripplanner.transit.model._data.PatternTestModel;
 import org.opentripplanner.transit.model._data.TransitModelForTest;
 import org.opentripplanner.transit.model.network.TripPattern;
 import org.opentripplanner.transit.model.site.RegularStop;
