+++ src/test/java/org/opentripplanner/transit/service/PatternByServiceDatesFilterTest.java
@@ -4,7 +4,7 @@ import static java.time.LocalDate.parse;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.opentripplanner.apis.gtfs.PatternTestModel.ROUTE_1;
+import static org.opentripplanner.transit.model._data.PatternTestModel.ROUTE_1;
 import static org.opentripplanner.transit.service.PatternByServiceDatesFilterTest.FilterExpectation.NOT_REMOVED;
 import static org.opentripplanner.transit.service.PatternByServiceDatesFilterTest.FilterExpectation.REMOVED;
 
@@ -13,8 +13,8 @@ import java.util.List;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.opentripplanner.apis.gtfs.PatternTestModel;
 import org.opentripplanner.apis.gtfs.model.LocalDateRange;
+import org.opentripplanner.transit.model._data.PatternTestModel;
 import org.opentripplanner.transit.model.network.TripPattern;
 
 class PatternByServiceDatesFilterTest {
